InventoryManagement/requirements.txt
asgiref
Django
sqlparse
psycopg2-binary
django-leaflet
django-guardian
